# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Math Functions API
  version: 1.0.0
  description: API для работы с математическими функциями
servers:
  - url: http://localhost:8080/api
    description: Local server
security:
  - basicAuth: []
paths:
...
  /points:
    get:
      summary: Получить точки функции в диапазоне X
      tags: [Points]
      parameters:
        - name: functionId
          in: query
          required: true
          schema:
            type: integer
        - name: xFrom
          in: query
          required: false
          schema:
            type: number
        - name: xTo
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PointResponse'
    post:
      summary: Добавить вычисленную точку
      tags: [Points]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                functionId:
                  type: integer
                xValue:
                  type: number
                yValue:
                  type: number
      responses:
        '201':
          description: Точка добавлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointResponse'
  /points/{id}:
    get:
      summary: Получить точку по ID
      tags: [Points]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointResponse'
        '404':
          description: Точка не найдена
    put:
      summary: Обновить точку
      tags: [Points]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                yValue:
                  type: number
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointResponse'
        '404':
          description: Точка не найдена
    delete:
      summary: Удалить точку
      tags: [Points]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Удалено
        '404':
          description: Точка не найдена
  /points/search:
    get:
      summary: Найти точку по координатам
      tags: [Points]
      parameters:
        - name: x
          in: query
          required: false
          schema:
            type: number
        - name: y
          in: query
          required: false
          schema:
            type: number
        - name: functionId
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PointResponse'

  # ОПЕРАЦИИ С ДОСТУПОМ
...
components:
  schemas:
...
    PointResponse:
      type: object
      properties:
        pointId:
          type: integer
        functionId:
          type: integer
        xValue:
          type: number
        yValue:
          type: number
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Basic authentication with username and password